requirements.txt
loguru>=0.5.2
websockets>=8.1
click>=7.1.2
aiofiles>=0.6.0
aiohttp>=3.7.3
requests>=2.25.1
attrs>=20.3.0
cattrs>=1.3.0
glom>=20.11.0
dependency-injector>=4.27.0
flask>=1.1.2
pandas>=1.2.3
Faker>=6.5.0
tabulate>=0.8.9
pymysql>=1.0.2
DBUtils>=2.0
cryptography>=3.4.6
selenium>=3.141.0
schedule>=1.0.0
grpcio>=1.36.1
protobuf>=3.15.6
grpcio-tools>=1.36.1
lxml~=4.6.3